Remove Sr. Helper Rank
The Sr. Helper rank seems to be dormant so this is listed as a separate commit.

The Sr. Helper rank could be used as a dedicate rank for those who wish to remain as a dedicate helper at which this commit should not merge.

The Helper permission config `helper.yml` already has the permissions  of Sr. Helper (such as checkmuted)